refactor sidebar navigate functionality
