+++ src/components/homePage/SidebarButton.jsx
@@ -1,26 +1,25 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useNavigate } from "react-router-dom";
-import useThemeStore from "../../stores/themeStore";
+import useSidebarStore from "../../stores/sidebarStore";
 import PropTypes from "prop-types";
+
 const SidebarButton = ({
     icon,
     text,
     textStyle,
     iconSize,
     iconColor,
-    onClick,
     style,
     path,
 }) => {
     const navigate = useNavigate();
+    const setTitle = useSidebarStore((state) => state.setTitle);
+    const closeSidebar = useSidebarStore((state) => state.closeSidebar);
     const handleClick = (e) => {
-        console.log({ text });
         e.stopPropagation();
-        if (onClick) {
-            onClick();
-        }
+        closeSidebar();
+        setTitle(text);
         navigate(path);
-        console.log(useThemeStore.getState().themeMode);
     };
     return (
         <button
@@ -50,7 +49,6 @@ SidebarButton.propTypes = {
     textStyle: PropTypes.string,
     iconSize: PropTypes.string,
     iconColor: PropTypes.string,
-    onClick: PropTypes.func,
     style: PropTypes.string,
     path: PropTypes.string,
 };
